Rename API app
